Split out MongoConnection into a separate common file.
